requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.12.2
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
loguru==0.7.3
lxml==4.9.3
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
nltk==3.9.1
numpy==1.26.4
openai==1.97.1
packaging==25.0
Pillow==10.1.0
pinecone-client==2.2.4
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.1.0
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.16.1
sentence-transformers==5.0.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.23
starlette==0.47.2
sympy==1.14.0
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.54.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
